+++ db/migrations/20181114200132_create_vow_storage_tables.sql
@@ -2,7 +2,7 @@
 CREATE TABLE maker.vow_vat
 (
     id        SERIAL PRIMARY KEY,
-    diff_id   INTEGER NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
+    diff_id   BIGINT  NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
     header_id INTEGER NOT NULL REFERENCES headers (id) ON DELETE CASCADE,
     vat       TEXT,
     UNIQUE (diff_id, header_id, vat)
@@ -11,7 +11,7 @@ CREATE TABLE maker.vow_vat
 CREATE TABLE maker.vow_flapper
 (
     id        SERIAL PRIMARY KEY,
-    diff_id   INTEGER NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
+    diff_id   BIGINT  NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
     header_id INTEGER NOT NULL REFERENCES headers (id) ON DELETE CASCADE,
     flapper   TEXT,
     UNIQUE (diff_id, header_id, flapper)
@@ -20,7 +20,7 @@ CREATE TABLE maker.vow_flapper
 CREATE TABLE maker.vow_flopper
 (
     id        SERIAL PRIMARY KEY,
-    diff_id   INTEGER NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
+    diff_id   BIGINT  NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
     header_id INTEGER NOT NULL REFERENCES headers (id) ON DELETE CASCADE,
     flopper   TEXT,
     UNIQUE (diff_id, header_id, flopper)
@@ -29,7 +29,7 @@ CREATE TABLE maker.vow_flopper
 CREATE TABLE maker.vow_sin_integer
 (
     id        SERIAL PRIMARY KEY,
-    diff_id   INTEGER NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
+    diff_id   BIGINT  NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
     header_id INTEGER NOT NULL REFERENCES headers (id) ON DELETE CASCADE,
     sin       numeric,
     UNIQUE (diff_id, header_id, sin)
@@ -38,7 +38,7 @@ CREATE TABLE maker.vow_sin_integer
 CREATE TABLE maker.vow_sin_mapping
 (
     id        SERIAL PRIMARY KEY,
-    diff_id   INTEGER NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
+    diff_id   BIGINT  NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
     header_id INTEGER NOT NULL REFERENCES headers (id) ON DELETE CASCADE,
     era       numeric,
     tab       numeric,
@@ -51,7 +51,7 @@ CREATE INDEX vow_sin_mapping_era_index
 CREATE TABLE maker.vow_ash
 (
     id        SERIAL PRIMARY KEY,
-    diff_id   INTEGER NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
+    diff_id   BIGINT  NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
     header_id INTEGER NOT NULL REFERENCES headers (id) ON DELETE CASCADE,
     ash       numeric,
     UNIQUE (diff_id, header_id, ash)
@@ -60,7 +60,7 @@ CREATE TABLE maker.vow_ash
 CREATE TABLE maker.vow_wait
 (
     id        SERIAL PRIMARY KEY,
-    diff_id   INTEGER NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
+    diff_id   BIGINT  NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
     header_id INTEGER NOT NULL REFERENCES headers (id) ON DELETE CASCADE,
     wait      numeric,
     UNIQUE (diff_id, header_id, wait)
@@ -69,7 +69,7 @@ CREATE TABLE maker.vow_wait
 CREATE TABLE maker.vow_dump
 (
     id        SERIAL PRIMARY KEY,
-    diff_id   INTEGER NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
+    diff_id   BIGINT  NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
     header_id INTEGER NOT NULL REFERENCES headers (id) ON DELETE CASCADE,
     dump      NUMERIC,
     UNIQUE (diff_id, header_id, dump)
@@ -78,7 +78,7 @@ CREATE TABLE maker.vow_dump
 CREATE TABLE maker.vow_sump
 (
     id        SERIAL PRIMARY KEY,
-    diff_id   INTEGER NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
+    diff_id   BIGINT  NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
     header_id INTEGER NOT NULL REFERENCES headers (id) ON DELETE CASCADE,
     sump      numeric,
     UNIQUE (diff_id, header_id, sump)
@@ -87,7 +87,7 @@ CREATE TABLE maker.vow_sump
 CREATE TABLE maker.vow_bump
 (
     id        SERIAL PRIMARY KEY,
-    diff_id   INTEGER NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
+    diff_id   BIGINT  NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
     header_id INTEGER NOT NULL REFERENCES headers (id) ON DELETE CASCADE,
     bump      numeric,
     UNIQUE (diff_id, header_id, bump)
@@ -96,7 +96,7 @@ CREATE TABLE maker.vow_bump
 CREATE TABLE maker.vow_hump
 (
     id        SERIAL PRIMARY KEY,
-    diff_id   INTEGER NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
+    diff_id   BIGINT  NOT NULL REFERENCES storage_diff (id) ON DELETE CASCADE,
     header_id INTEGER NOT NULL REFERENCES headers (id) ON DELETE CASCADE,
     hump      numeric,
     UNIQUE (diff_id, header_id, hump)
